+++ example/linux/flutter/ephemeral/.plugin_symlinks/device_info_plus/example/android/app/src/main/kotlin/io/flutter/plugins/deviceinfoexample/example/MainActivity.kt
@@ -1,5 +1,0 @@
-package io.flutter.plugins.deviceinfoexample.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()