+++ example/linux/flutter/ephemeral/.plugin_symlinks/device_info_plus/lib/src/device_info_plus_macos.dart
@@ -1,1 +1,0 @@
-/// The platform interface handles the method channel calls. This file exists to silence errors on pub.