+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.4
 
-* 
+* Update dependencies
 
 ## 0.0.3