+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.5
+
+* Add support for new iPhones: 16, 16 Plus, 16 Pro, 16 Pro Max
+* Refactor internals
+* Update README
+* Update dependencies
+* Add older Flutter versions support
+* Add device orientation check
+
 ## 0.0.4
 
 * Update dependencies